Render wire integer types as TS `number`, not `bigint`
Every transferred integer in the GridFPV wire contract is bounded far below
`Number.MAX_SAFE_INTEGER` (~9e15) in our domain, and serde already serialises
them as JSON numbers. Rendering them as TS `bigint` only produced wire-vs-type
mismatches (a `number` arriving where a `bigint` was declared). This keeps the
Rust types as `i64`/`u64` (the engine still does arithmetic on them) and changes
only the ts-rs TS rendering plus the now-pointless bigint plumbing in the
frontend.

ts-rs rendering (→ `number`):
- `SourceTime`, `LogRef`, `Cursor`: transparent newtypes, `#[ts(as = "f64")]`.
- `Pass.sequence`, `Penalty::TimeAdded.micros`, `Lap.duration_micros`,
  `WinCondition::Timed.window_micros`, `PilotProgress.last_lap_micros`:
  `#[ts(type = "number")]`.
- `Metric::BestLapMicros` / `BestConsecutiveMicros`:
  `#[ts(type = "number | null")]`.
- Updated `Cursor` doc comment (no longer claims u64 exceeds JS safe range).
- Regenerated `bindings/`; zero `bigint` remain.

Frontend:
- protocol-client: stripped `toCursor`/`stringifyWire`/`normalize*` (cursors are
  plain numbers from JSON now); `streamSeq`/`cursor` are `number`; `0n`→`0`,
  `+ 1n`→`+ 1`. Ordering/apply logic unchanged. Updated tests to plain numbers.
- components `formatMicros`: param is now `number | null | undefined`; kept the
  `Number()` coercion as belt-and-suspenders; updated doc comment.
- rd-console consumers (setup/marshaling/control/results + screens + fixtures/
  tests) updated to build/assert plain `number`s instead of bigint.

Part of #13 (v0.4 wire data-type contract).

+++ frontend/apps/rd-console/src/lib/marshaling.ts
@@ -10,7 +10,8 @@
  *
  * `Command`'s five marshaling variants are pure functions of their inputs here, so the
  * screen builds them declaratively and the tests assert the exact wire shape. Targets
- * are `LogRef` (a u64 log offset, `bigint`); times are `SourceTime` (µs, `bigint`).
+ * are `LogRef` (a u64 log offset, rendered as a TS `number`); times are `SourceTime`
+ * (µs, rendered as a TS `number`).
  */
 
 import type {
@@ -58,7 +59,7 @@ export function applyPenaltyCommand(
 
 /** Build a `TimeAdded` penalty from a whole-second amount (the console's input unit). */
 export function timeAddedPenalty(seconds: number): Penalty {
-  return { TimeAdded: { micros: BigInt(Math.round(seconds * 1_000_000)) } };
+  return { TimeAdded: { micros: Math.round(seconds * 1_000_000) } };
 }
 
 /** The disqualification penalty. */
@@ -66,5 +67,5 @@ export const DISQUALIFY: Penalty = 'Disqualify';
 
 /** Convert the console's whole-second time input to `SourceTime` microseconds. */
 export function secondsToSourceTime(seconds: number): SourceTime {
-  return BigInt(Math.round(seconds * 1_000_000));
+  return Math.round(seconds * 1_000_000);
 }
